Commits character map and deletes previous enigma file
